# perf(middleware/logger): optimize color status