--- a/src/middleware/logger/index.ts
+++ b/src/middleware/logger/index.ts
@@ -32,18 +32,14 @@ const colorStatus = (status: number) => {
     return `${status}`
   }
   switch ((status / 100) | 0) {
-    case 7: // magenta = ???
-      return `\x1b[35m${status}\x1b[0m`
     case 5: // red = error
       return `\x1b[31m${status}\x1b[0m`
+    case 4: // yellow = warning
+      return `\x1b[33m${status}\x1b[0m`
     case 3: // cyan = redirect
       return `\x1b[36m${status}\x1b[0m`
-    case 2:
-    case 1: // green = success
+    case 2: // green = success
       return `\x1b[32m${status}\x1b[0m`
-    case 4:
-    case 0: // yellow = warning
-      return `\x1b[33m${status}\x1b[0m`
   }
 }