--- a/src/middleware/logger/index.ts
+++ b/src/middleware/logger/index.ts
@@ -28,19 +28,22 @@ const time = (start: number) => {
 
 const colorStatus = (status: number) => {
   const colorEnabled = getColorEnabled()
-  if (!colorEnabled) {
-    return `${status}`
+  if (colorEnabled) {
+    switch ((status / 100) | 0) {
+      case 5: // red = error
+        return `\x1b[31m${status}\x1b[0m`
+      case 4: // yellow = warning
+        return `\x1b[33m${status}\x1b[0m`
+      case 3: // cyan = redirect
+        return `\x1b[36m${status}\x1b[0m`
+      case 2: // green = success
+        return `\x1b[32m${status}\x1b[0m`
+    }
   }
-  switch ((status / 100) | 0) {
-    case 5: // red = error
-      return `\x1b[31m${status}\x1b[0m`
-    case 4: // yellow = warning
-      return `\x1b[33m${status}\x1b[0m`
-    case 3: // cyan = redirect
-      return `\x1b[36m${status}\x1b[0m`
-    case 2: // green = success
-      return `\x1b[32m${status}\x1b[0m`
-  }
+  // Fallback to unsupported status code.
+  // E.g.) Bun and Deno supports new Response with 101, but Node.js does not.
+  // And those may evolve to accept more status.
+  return `${status}`
 }
 
 type PrintFunc = (str: string, ...rest: string[]) => void