@yusukebe 2024-11-01
@exoego 

Thanks!

> I've noticed some status codes (e.g. `700`) are unreachable, so deleted.

You are right. Good point!